crypto: msm: Fix size of descriptor buffer
When the number of entries in the scatter gather list is larger
than 64, crypto operation fail.  This is due to incorrect use
of macro to allocate memory for the descriptor size.

Fix is to replace use of QCE_SIZE_BAM_DSCR with the size of the
sps_iovec data structure. Further, more use the correct macro when
reporting the size of the descriptor to BAM.

Change-Id: I4396ad8407b6b91d7a34367a2777cb920f81457b
Signed-off-by: Mona Hossain <mhossain@codeaurora.org>