misc: pmem: Support for dynamic allocation through CMA
A new member, named use_cma, has been added to the struct android_pmem_platform_data.
If use_cma is set to 1, the memory amount of PMEM heap will be dynamically
reserved and then released when it's no more needed through CMA api.
This solution is ready to be used with bitmap allocator.